Add support for packing Python objects to primitive values
Sometimes it's appropriate for the JS representation of an object to be a plain JSON-serializable value such as a string, rather than a custom-defined object. This would be possible by registering a subclass of StringAdapter / BaseAdapter, except `register` has over-zealous validation that rejects anything other than object-based Adapters. Relax this validation and add some documentation to make this use-case official.